requirements.txt
streamlit==1.31.0
openai==1.12.0
plotly==5.18.0
pandas==2.2.0
python-dotenv==1.0.1
sqlalchemy==2.0.25
pyodbc==5.0.1
psycopg2-binary==2.9.9
pymysql==1.1.0
transformers==4.36.0
torch>=2.0.0
accelerate==0.25.0
sentencepiece==0.1.99
protobuf==3.20.3